Split barcode to image logic into a separate module
