+++ src/labelle/lib/barcode_writer.py
@@ -6,10 +6,9 @@
 # this notice are preserved.
 # === END LICENSE STATEMENT ===
 
-from typing import List, NamedTuple, NewType, Tuple
+from typing import List, NamedTuple, NewType
 
 from barcode.writer import BaseWriter
-from PIL import Image, ImageDraw
 
 BinaryString = NewType("BinaryString", str)
 """A string that's been validated to contain only '0's and '1's."""
@@ -19,43 +18,6 @@ def _validate_string_as_binary(s: str) -> BinaryString:
     if not all(c in ("0", "1") for c in s):
         raise ValueError("Barcode can only contain 0 and 1")
     return BinaryString(s)
-
-
-def _mm2px(mm: float, dpi: float = 25.4) -> float:
-    return (mm * dpi) / 25.4
-
-
-def _list_of_runs(line: BinaryString) -> List[int]:
-    # Pack line to list give better gfx result, otherwise in can
-    # result in aliasing gaps
-    # '11010111' -> [2, -1, 1, -1, 3]
-    c = 1
-    mlist = []
-    for i in range(0, len(line)):
-        if i + 1 < len(line) and line[i] == line[i + 1]:
-            c += 1
-        else:
-            if line[i] == "1":
-                mlist.append(c)
-            else:
-                mlist.append(-c)
-            c = 1
-    return mlist
-
-
-def _calculate_size(
-    *,
-    modules_per_line: int,
-    number_of_lines: int,
-    quiet_zone: float,
-    module_width: float,
-    module_height: float,
-    vertical_margin: float,
-    dpi: float = 25.4,
-) -> Tuple[int, int]:
-    width = 2 * quiet_zone + modules_per_line * module_width
-    height = vertical_margin * 2 + module_height * number_of_lines
-    return int(_mm2px(width, dpi)), int(_mm2px(height, dpi))
 
 
 class BarcodeResult(NamedTuple):
@@ -70,76 +32,3 @@ class SimpleBarcodeWriter(BaseWriter):
             raise ValueError("Barcode expected to have only one line")
         line = _validate_string_as_binary(code[0])
         return BarcodeResult(line=line, quiet_zone=self.quiet_zone)
-
-
-def convert_binary_string_to_barcode_image(
-    line: BinaryString, quiet_zone: float, module_height: float
-) -> Image.Image:
-    """Render a barcode string into an image.
-
-    line: A string of 0s and 1s representing the barcode.
-    """
-    module_width = 2
-    background = "black"
-    foreground = "white"
-    vertical_margin = 8
-    dpi = 25.4
-
-    width, height = _calculate_size(
-        modules_per_line=len(line),
-        number_of_lines=1,
-        dpi=dpi,
-        quiet_zone=quiet_zone,
-        module_width=module_width,
-        module_height=module_height,
-        vertical_margin=vertical_margin,
-    )
-    image = Image.new("1", (width, height), background)
-    draw = ImageDraw.Draw(image)
-
-    ypos = vertical_margin
-    mlist = _list_of_runs(line)
-    # Left quiet zone is x startposition
-    xpos = quiet_zone
-    for mod in mlist:
-        if mod < 1:
-            color = background
-        else:
-            color = foreground
-        # remove painting for background colored tiles?
-        _paint_module(
-            xpos=xpos,
-            ypos=ypos,
-            width=module_width * abs(mod),
-            color=color,
-            dpi=dpi,
-            module_height=module_height,
-            draw=draw,
-        )
-        xpos += module_width * abs(mod)
-    return _finish(image)
-
-
-def _paint_module(
-    *,
-    xpos: float,
-    ypos: float,
-    width: float,
-    color: str,
-    dpi: float,
-    module_height: float,
-    draw: ImageDraw.ImageDraw,
-) -> None:
-    size = (
-        (_mm2px(xpos, dpi), _mm2px(ypos, dpi)),
-        (
-            _mm2px(xpos + width, dpi),
-            _mm2px(ypos + module_height, dpi),
-        ),
-    )
-    draw.rectangle(size, outline=color, fill=color)
-
-
-def _finish(image: Image.Image) -> Image.Image:
-    # although Image mode set to "1", draw function writes white as 255
-    return image.point(lambda x: 1 if x > 0 else 0, mode="1")
